fix(threads,composition): pre-limit tombstone count, module split, ref-replay assertions
- Tombstone decision in list_deferred_busy_messages now uses the untruncated
  match count in both backends; limit: Some(0) can no longer evict the
  presence marker while live DeferredBusy rows exist (contract tests added)
- contract.rs request docs no longer claim drain resubmits via the inbound
  path — it replays persisted canonical refs through the coordinator
- deferred_busy_drain split into mod.rs (494-line observer) + tests.rs;
  no-op tests/deferred_busy_drain_e2e.rs target deleted
- New Scenario L captures the drained SubmitTurnRequest and asserts the
  persisted webui-src/webui-reply refs are replayed verbatim
